backend/requirements.txt
# Core API
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sse-starlette>=1.8.2
python-multipart>=0.0.9

# LangGraph & LangChain
langgraph>=0.1.5
langchain>=0.2.0
langchain-google-genai>=1.0.7
langchain-core>=0.2.0

# Google Gemini
google-generativeai>=0.7.0

# Supabase
supabase>=2.4.0

# Vector Search
faiss-cpu>=1.8.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.0

# PDF Parsing
pdfplumber>=0.11.0
fpdf2>=2.7.9

# Fuzzy Matching
thefuzz>=0.22.1
python-Levenshtein>=0.25.1

# Retry / Failure Handling
tenacity>=8.3.0

# Config
python-dotenv>=1.0.1

# HTTP
httpx>=0.27.0
